% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Anatolia_EF_CLR}
\alias{Anatolia_EF_CLR}
\title{Anatolia_EF_CLR}
\format{
A data frame with gene expression values.
}
\source{
Internal example dataset for testing.
}
\description{
Example data: Gene expression factors from Anatolia population.
}
