% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecation.R
\name{.deprecate_function}
\alias{.deprecate_function}
\title{Deprecation helpers for polylinkR tidyverse naming transition}
\usage{
.deprecate_function(old_name, new_name, ...)
}
\arguments{
\item{old_name}{Character string of the deprecated function name}

\item{new_name}{Character string of the new function name}

\item{...}{Arguments passed to the new function}
}
\value{
Result from calling the new function
}
\description{
Internal helper functions for managing the transition from legacy naming
conventions to tidyverse-style naming in polylinkR v0.6.0. These functions
provide soft deprecation with informative warning messages.
}
\examples{
\dontrun{
# Old function definition:
plR_read <- function(...) {
  .deprecate_function("plR_read", "read_polylinkr_data", ...)
}
}

}
\keyword{Creates}
\keyword{Deprecate}
\keyword{a}
\keyword{alias}
\keyword{an}
\keyword{calls}
\keyword{deprecated}
\keyword{for}
\keyword{function}
\keyword{informative}
\keyword{internal}
\keyword{name}
\keyword{new}
\keyword{that}
\keyword{the}
\keyword{warning.}
\keyword{with}
