% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecation.R
\name{.deprecate_param}
\alias{.deprecate_param}
\title{Deprecate a parameter name}
\usage{
.deprecate_param(arg_list, old_param, new_param)
}
\arguments{
\item{arg_list}{List of arguments from the calling function}

\item{old_param}{Character string of the deprecated parameter name}

\item{new_param}{Character string of the new parameter name}
}
\value{
The value for the parameter (from old or new name), or NULL
}
\description{
Checks if a deprecated parameter name was used and issues a warning,
returning the value mapped to the new parameter name.
}
\examples{
\dontrun{
# In function definition:
read_polylinkr_data <- function(input_path = NULL, input.path = NULL) {
  # Handle deprecated parameter
  if (!is.null(input.path) && is.null(input_path)) {
    input_path <- .deprecate_param(
      list(input.path = input.path, input_path = input_path),
      "input.path", "input_path"
    )
  }
  # ... rest of function
}
}

}
\keyword{internal}
