% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecation.R
\name{.make_deprecated_alias}
\alias{.make_deprecated_alias}
\title{Create a deprecated function alias}
\usage{
.make_deprecated_alias(old_name, new_name, new_fun)
}
\arguments{
\item{old_name}{Character string - name of the deprecated function}

\item{new_name}{Character string - name of the new function}

\item{new_fun}{Function object - the new function to wrap}
}
\value{
A function that calls the new function with deprecation warning
}
\description{
Helper to create a deprecated wrapper around a new function.
This is used to maintain backward compatibility while warning users.
}
\keyword{internal}
