% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecation.R
\name{.map_deprecated_params}
\alias{.map_deprecated_params}
\title{Map deprecated parameter names to new names}
\usage{
.map_deprecated_params(arg_list, param_map)
}
\arguments{
\item{arg_list}{List of arguments (usually from match.call or list(...))}

\item{param_map}{Named character vector where names are old parameters
and values are new parameters}
}
\value{
Modified argument list with deprecated parameters renamed
}
\description{
Takes a list of arguments and maps any deprecated parameter names
to their new equivalents, issuing deprecation warnings.
}
\examples{
\dontrun{
# Define parameter mapping
param_map <- c(
  "input.path" = "input_path",
  "obj.info.path" = "object_info_path",
  "n.perm" = "n_permutations"
)

# In function:
args <- .map_deprecated_params(args, param_map)
}

}
\keyword{internal}
