% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{plR_permute}
\alias{plR_permute}
\title{Gene set enrichment with control for confounding covariates (deprecated)}
\usage{
plR_permute(
  plR.input,
  permute = TRUE,
  n.perm = 500000L,
  n.boot = 30L,
  alt = "upper",
  md.meth = "robust",
  kern.wt.max = 0.05,
  kern.bound = "auto",
  kern.func = "normal",
  kern.scale = "auto",
  gpd.cutoff = 5000L/n.perm,
  seed = NULL,
  verbose = TRUE,
  n.cores = "auto",
  fut.plan = "auto"
)
}
\arguments{
\item{plR.input}{Deprecated. Use \code{plr_input}.}

\item{permute}{Deprecated. Use \code{permute}.}

\item{n.perm}{Deprecated. Use \code{n_permutations}.}

\item{n.boot}{Deprecated. Use \code{n_bootstraps}.}

\item{alt}{Deprecated. Use \code{alternative}.}

\item{md.meth}{Deprecated. Use \code{md_method}.}

\item{kern.wt.max}{Deprecated. Use \code{kernel_weight_max}.}

\item{kern.bound}{Deprecated. Use \code{kernel_boundary}.}

\item{kern.func}{Deprecated. Use \code{kernel_function}.}

\item{kern.scale}{Deprecated. Use \code{kernel_scale}.}

\item{gpd.cutoff}{Deprecated. Use \code{gpd_cutoff}.}

\item{seed}{Deprecated. Use \code{seed}.}

\item{verbose}{Deprecated. Use \code{verbose}.}

\item{n.cores}{Deprecated. Use \code{n_cores}.}

\item{fut.plan}{Deprecated. Use \code{future_plan}.}
}
\description{
This function is deprecated. Please use \code{permute_polylinkr_data()} instead.
}
