% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{plR_rescale}
\alias{plR_rescale}
\title{Gene set enrichment on scores rescaled for genetic autocorrelation (deprecated)}
\usage{
plR_rescale(
  plR.input,
  rescale = TRUE,
  fast = TRUE,
  ac = NULL,
  cgm.bin = 30,
  cgm.range = "auto",
  cgm.wt.max = 0.05,
  emp.bayes = "auto",
  min.rho = 1e-05,
  verbose = TRUE,
  n.cores = "auto",
  fut.plan = "auto"
)
}
\arguments{
\item{plR.input}{Deprecated. Use \code{plr_input}.}

\item{rescale}{Deprecated. Use \code{rescale}.}

\item{fast}{Deprecated. Use \code{fast}.}

\item{ac}{Deprecated. Use \code{ac}.}

\item{cgm.bin}{Deprecated. Use \code{cgm_bin}.}

\item{cgm.range}{Deprecated. Use \code{cgm_range}.}

\item{cgm.wt.max}{Deprecated. Use \code{cgm_weight_max}.}

\item{emp.bayes}{Deprecated. Use \code{empirical_bayes}.}

\item{min.rho}{Deprecated. Use \code{min_rho}.}

\item{verbose}{Deprecated. Use \code{verbose}.}

\item{n.cores}{Deprecated. Use \code{n_cores}.}

\item{fut.plan}{Deprecated. Use \code{future_plan}.}
}
\description{
This function is deprecated. Please use \code{rescale_polylinkr_data()} instead.
}
