% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polylinkR-package.R
\docType{package}
\name{polylinkR-package}
\alias{polylinkR}
\alias{polylinkR-package}
\title{polylinkR: Polygenic Pathway Enrichment with Linkage Structure}
\description{
Gene-based pathway enrichment that accounts for linkage disequilibrium amongst loci, using a permutation approach that preserves linkage structure. This version is under active development and must not be used as the sole basis for production or clinical decisions without independent validation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ACAD-UofA/PolyLinkR}
  \item \url{https://acad-uofa.github.io/PolyLinkR/}
  \item Report bugs at \url{https://github.com/ACAD-UofA/PolyLinkR/issues}
}

}
\author{
\strong{Maintainer}: PolyLinkR authors \email{polylinkr@acad.edu.au}

Other contributors:
\itemize{
  \item ACAD University of Adelaide [copyright holder, funder]
}

}
\keyword{internal}
