% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polylinkr.R
\name{polylinkr}
\alias{polylinkr}
\title{Compute null distribution and probability estimation of gene score enrichment in pathways or gene sets.}
\usage{
polylinkr(
  set.info,
  obj.info,
  set.obj,
  n.cores = NA,
  emp.nruns = 10000,
  NN = 1000
)
}
\arguments{
\item{set.info}{data.frame: four required fields; gene IDs, associated scores, chromosome, and start position}

\item{obj.info}{data.frame: two required fields; genomic regions and associated scores}

\item{set.obj}{data.frame: two required fields; genomic regions and associated scores}

\item{n.cores}{integer: number of cores to run in parallel}

\item{emp.nruns}{integer: number of iterations used to compute null}

\item{NN}{integer: subset of interations for summary statistic calculations; default = 1000}
}
\value{
Returns a data.table.
}
\description{
This is the core function of PolyLinkR, it runs the permuation algorithm that
creates a random mapping of genomic scores while preserving the inherent linkage
disequilibrium amongt the different genomic regions.
This process performed iteratively to generate a null distribution for testing
enrichment in biological pathways or gene sets.
}
\examples{
output = polylinkr(obj.info = Anatolia_EF_CLR, set.info = PolyLinkR_SetInfo, set.obj = PolyLinkR_SetObj,
             n.cores = 8, emp.nruns = 1000, NN = 10000)

}
\seealso{
\code{\link{PolyLinkR_SetInfo}}, \code{\link{PolyLinkR_SetObj}}
}
