% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polylinkr_set_info}
\alias{polylinkr_set_info}
\title{PolylinkR example gene set information}
\format{
A data frame with setID, setName, setSource columns.
}
\source{
Internal example dataset for testing.
}
\usage{
data(polylinkr_set_info)
}
\description{
Example gene set information for testing.
}
\keyword{datasets}
