% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polylinkr_set_obj}
\alias{polylinkr_set_obj}
\title{PolylinkR example gene-to-set mapping}
\format{
A data frame with setID and objID columns.
}
\source{
Internal example dataset for testing.
}
\usage{
data(polylinkr_set_obj)
}
\description{
Example gene-to-set mapping for testing.
}
\keyword{datasets}
