% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{PolyLinkR_SetInfo}
\alias{PolyLinkR_SetInfo}
\title{PolyLinkR_SetInfo}
\format{
A data frame with setID, setName, setSource columns.
}
\source{
Internal example dataset for testing.
}
\description{
Example gene set information for testing.
}
