% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{PolyLinkR_SetObj}
\alias{PolyLinkR_SetObj}
\title{PolyLinkR_SetObj}
\format{
A data frame with setID and objID columns.
}
\source{
Internal example dataset for testing.
}
\description{
Example gene-to-set mapping for testing.
}
