% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_plr.R
\name{print.plR}
\alias{print.plR}
\title{Print method for \code{plR} objects}
\usage{
\method{print}{plR}(x, ...)
}
\arguments{
\item{x}{\code{plR} class object; typically output from a \code{polylinkR}
function.}

\item{...}{Additional arguments passed to \code{print}.}
}
\value{
An overview of the \code{plR} object.
}
\description{
Prints an overview of a \code{plR} object, including its processing
history and available contents.
}
\examples{
\dontrun{
# Provide a snapshot of core files (assuming `my_plR` is a valid plR object)
print(my_plR)

# Simply typing the object name also prints the snapshot
my_plR
}
}
\keyword{internal}
