% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_polylinkr_data}
\alias{prune_polylinkr_data}
\title{Gene set enrichment adjusting for multiple testing and shared genes}
\usage{
prune_polylinkr_data(
  plr_input,
  n_fdr = 300L,
  estimate_pi0 = TRUE,
  tolerance = 0.001,
  verbose = TRUE,
  n_cores = "auto",
  future_plan = "auto"
)
}
\arguments{
\item{plr_input}{\code{plR} class object; output from \code{polylinkR::permute_polylinkr_data} or
\code{polylinkR::rescale_polylinkr_data}. Required.}

\item{n_fdr}{\code{integer}; number of times to replicate the pruning procedure.
Used to estimate FDR-corrected p (q) values using a histogram method.
Defaults to \code{300L}. Range \code{[100, Inf)} and must be exactly
divisible by \code{100}.}

\item{estimate_pi0}{\code{logical}; should \code{pi0} be estimated during FDR correction?
Defaults to \code{TRUE}. If \code{FALSE}, \code{pi0} is set to \code{1}.}

\item{tolerance}{\code{numeric}; minimum difference between successive \code{pi0}
estimates for estimation to stop. Defaults to \code{1e-3}. Range
\code{(0, 0.01]}.}

\item{verbose}{\code{logical}; should progress reporting be enabled? Defaults to
\code{TRUE}.}

\item{n_cores}{\code{integer}; number of cores for parallel processing. Defaults to
\code{1} or \code{maximum cores - 1}. Must be in \code{[1, maximum cores]}.}

\item{future_plan}{\code{character}; parallel backend from the \code{future} package.
Defaults to user \code{n_cores} choice or checks cores, choosing
\code{"sequential"} on single-core and \code{"multisession"} on
multi-core systems. Options: \code{"multisession"},
\code{"multicore"}, \code{"cluster"}, \code{"sequential"}.}
}
\value{
A \code{plR} S3 object containing:
\describe{
\item{\code{obj.info}}{\code{data.table} for each gene (object).}
\item{\code{set.info}}{\code{data.table} for each gene set.}
\item{\code{set.obj}}{\code{data.table} mapping genes to gene sets.}
}

Pruned set scores (i.e., corrected for between–gene-set correlations
caused by shared genes) are recorded in \code{set.info} as
\code{setScore.pr}. Corrected scores are shown in \code{setScore.pr.p}.
Results from multiple testing correction are shown in \code{setScore.pr.q}.

All \code{plR} S3 objects include auxiliary information and summaries as
attributes:
\describe{
\item{\code{plr.data}}{
Reusable datasets and parameters, including GPD fitting results and
autocovariance estimation.
}
\item{\code{plr.args}}{
Argument settings used in the function.
}
\item{\code{plr.session}}{
R session information and function run time.
}
\item{\code{plr.track}}{
Internal tracking number indicating functional steps.
}
\item{\code{class}}{
S3 object class (i.e., \code{plr}).
}
}

Each attribute can be accessed using \code{attr()} or \code{attributes()}.
The first four attribute classes aggregate information over successive
functions. For example, to access the \code{plr.data} attribute for the
\code{plR} object output after running \code{plR_permute}, use
\code{attributes(X)$plr_data$permute.data}, where \code{X} is the object
name. Similarly, the arguments used in \code{plR_read} are in
\code{attributes(X)$plr_args$read.args}.

The primary data structure of the \code{plR} object can be accessed using
\code{print()} or by simply typing the object's name.
}
\description{
Adjusts gene set scores to account dependencies between gene sets using a
pruning routine involving sequential p-value ranking and removal of shared
genes. Performs multiple testing correction on the pruned gene set p-values.
Note that this step require that at more than one gene set with shared genes
is present, otherwise the function will exit with an error message.
}
\examples{
\dontrun{
# Assuming `my_plr` is the result of `polylinkR::permute_polylinkr_data` or
# `polylinkR::rescale_polylinkr_data`

# Example 1: Basic usage
new_plr <- prune_polylinkr_data(plr_input = my_plr)

# Example 2: Increase iterations for null p-value distribution
new_plr <- prune_polylinkr_data(
   plr_input = my_plr,
   n_fdr = 1000L
)

# Example 3: Do not estimate pi0 (sets pi0 = 1)
new_plr <- prune_polylinkr_data(
   plr_input = my_plr,
   estimate_pi0 = FALSE
)
}
}
