% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_polylinkr_data}
\alias{read_polylinkr_data}
\title{Read and validate required files for polylinkR}
\usage{
read_polylinkr_data(
  input_path = NULL,
  object_info_path = NULL,
  gene_set_info_path = NULL,
  gene_set_mapping_path = NULL,
  variant_info_path = NULL,
  recombination_rate_path = NULL,
  min_set_size = 2L,
  max_set_size = Inf,
  group_label = NULL,
  mapping_function = "kosambi",
  object_buffer = "auto",
  object_statistic_function = "non.param",
  bin_size = 250L,
  objects_to_include = NULL,
  objects_to_exclude = NULL,
  sets_to_include = NULL,
  sets_to_exclude = NULL,
  merge_threshold = 0.95,
  remove_duplicate_genes = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{input_path}{\code{character}; path to the directory with input files.
Defaults to \code{NULL}. Not required if using separate file paths.
Automatically searches for required files with the following labels:
\itemize{
\item \code{setinfo}: a data frame with set-level information.
\item \code{objinfo}: a data frame with object-level information.
\item \code{setobj}: a data frame mapping objects to sets.
}
It also searches for optional files:
\itemize{
\item \code{recrate}: an optional file for genetic coordinate
conversion.
\item \code{varinfo}: an optional file for gene score generation.
}
Allowable file names include capitalisation of each word in the label
and use of internal separators (e.g., \code{set.info}, \code{set_info},
\code{SetInfo}, \code{Set.Info}, \code{Set_Info}). Also see the section
on required and optional input files for the columns needed in each
input file.}

\item{object_info_path}{\code{character}; path to the object info file. Defaults to
\code{NULL}. Ignored if \code{input_path} is provided; otherwise all
required file paths must be specified.}

\item{gene_set_info_path}{\code{character}; path to the gene set info file. Defaults to
\code{NULL}. Ignored if \code{input_path} is provided; otherwise all
required file paths must be specified.}

\item{gene_set_mapping_path}{\code{character}; path to the gene set mapping file. Defaults to
\code{NULL}. Ignored if \code{input_path} is provided; otherwise all
required file paths must be specified.}

\item{variant_info_path}{\code{character}; path to the variant info file. Defaults to
\code{NULL}. Optional file used for gene score generation. Ignored if
\code{input_path} is provided.}

\item{recombination_rate_path}{\code{character}; path to the recombination rate file. Defaults to
\code{NULL}. Optional file used for genetic coordinate conversion.
Ignored if \code{input_path} is provided.}

\item{min_set_size}{\code{integer}; minimum size of gene sets to be retained. Defaults to
\code{2}. Must be in the range \code{[2L, max_set_size)}.}

\item{max_set_size}{\code{integer}; maximum size of gene sets to be retained. Defaults to
\code{Inf}. Must be in the range \code{(min_set_size, Inf)}.}

\item{group_label}{\code{character}; label used to identify input files within a directory.
Defaults to \code{NULL}.}

\item{mapping_function}{\code{character}; mapping function to convert physical to genetic
distances. Options are \code{"Haldane"}, \code{"Kosambi"} (default),
\code{"Carter-Falconer"}, and \code{"Morgan"}.}

\item{object_buffer}{\code{numeric}; interval around genes (in base pairs) to include when
assigning values from the variant info file. Defaults to \code{1e4} if
variant info is provided and user does not set a value, otherwise it is
set to 0 if score assignment is not performed. User values must be in the
range \code{[0, 1e5L]}. Note that if the user provides their own gene
scores in the object info input file (i.e. not computed from the
variant info file), then the start and end positions must include any
buffer used to bin scores, otherwise polylinkR deconfounding and
autocorrelation inference will not be performed appropriately.}

\item{object_statistic_function}{\code{character}; function used to correct maximum gene scores based on
the number of overlapping summary statistics (SNPs or windows). Default is
\code{non.param}, a robust non-parametric method that uses binned data to
calculate median and median absolute deviation (MAD) to normalise scores.
Alternatively, \code{lm.logN} applies a linear regression to the log-transformed
SNP / bin counts (assumes a roughly linear relationship is appropriate). In
both cases, expected scores are estimated and gene scores calculated as the
residual value. Ignored if no variant info file is provided.}

\item{bin_size}{\code{integer}; gene set size interval for non-parametric correction.
Defaults to \code{250L}. Must be in the range \code{[50L, 1e3L]};
ignored if the parametric function is used.}

\item{objects_to_include}{\code{character} or \code{numeric} vector; object IDs of genes to
explicitly retain. Defaults to \code{NULL}.}

\item{objects_to_exclude}{\code{character} or \code{numeric} vector; object IDs of genes to
explicitly remove. Defaults to \code{NULL}.}

\item{sets_to_include}{\code{character} or \code{numeric} vector; set IDs of gene sets
to explicitly retain. Defaults to \code{NULL}.}

\item{sets_to_exclude}{\code{character} or \code{numeric} vector; set IDs of gene sets
to explicitly remove. Defaults to \code{NULL}.}

\item{merge_threshold}{\code{numeric}; minimum proportion of shared genes for merging gene sets.
Defaults to \code{0.95}. Must be in the range \code{(0, 1]}.}

\item{remove_duplicate_genes}{\code{logical}; should genes with identical genomic positions be
removed? Defaults to \code{FALSE}.}

\item{verbose}{\code{logical}; should progress messages be printed to the console?
Defaults to \code{TRUE}.}
}
\value{
A \code{plR} S3 object containing three complementary datasets:
\describe{
\item{\code{obj.info}}{
\code{data.table} containing information for each gene (object).
}
\item{\code{set.info}}{
\code{data.table} containing information for each gene set.
}
\item{\code{set.obj}}{
\code{data.table} containing the mapping of genes to gene sets.
}
}

All \code{plR} S3 objects include auxiliary information and summaries as
attributes:
\describe{
\item{\code{plr.data}}{
Reusable datasets and parameters, including GPD fitting results and
autocovariance estimation.
}
\item{\code{plr.args}}{
Argument settings used in the function.
}
\item{\code{plr.session}}{
R session information and function run time.
}
\item{\code{plr.track}}{
Internal tracking number indicating functional steps.
}
\item{\code{class}}{
S3 object class (i.e., \code{plr}).
}
}

Each attribute can be accessed using \code{attr()} or \code{attributes()}.
The first four attribute classes aggregate information over successive
functions. For example, to access the \code{plr.data} attribute for the
\code{plR} object output after running \code{plR_permute}, use
\code{attributes(X)$plR.data$permute.data}, where \code{X} is the object
name. Similarly, the arguments used in \code{plR_read} are in
\code{attributes(X)$plR.args$read.args}.

The primary data structure of the \code{plR} object can be accessed using
\code{print()} or by simply typing the object's name.
}
\description{
Reads and validates the input files for a \code{polylinkR} analysis. Handles
file path resolution, column name canonicalisation, data validation, and
optional filtering and data generation tasks like oordinate conversion or
gene score calculation.
}
\section{Required input file structure}{

The following three files are compulsory for a \code{polylinkR} analysis
and extend the format introduced in Polysel. All files must be comma-
separated (\code{.csv}) or tab-separated (\code{.tsv}), with a header
(see examples at \url{https://github.com/CMPG/polysel/tree/master/data}).

\describe{
\item{\code{set.info}}{
A \code{data.table} (and \code{data.frame}) with gene set information.
\itemize{
\item \strong{setID}: \code{character} or \code{factor} vector of
unique gene set identifiers. Required.
}
}
\item{\code{obj.info}}{
A \code{data.table} (and \code{data.frame}) with gene (objects)
information.
\itemize{
\item \strong{objID}: \code{character} or \code{factor} vector of
unique gene identifiers. Required.
\item \strong{objStat}: optional \code{numeric} vector of
pre-computed gene scores. Required if \code{var.info} is absent;
otherwise computed from \code{var.info}.
\item \strong{CovX}: \code{numeric} vector of covariate scores,
where \code{X} is a positive integer denoting covariate number.
Required for deconfounding gene scores (\code{objStat}) in
\code{plR_permute}.
\item \strong{chr}: \code{character} or \code{numeric} chromosome /
contig labels. Required for gene score deconfounding and gene set
score decorrelation in and \code{plR_permute} and \code{plR_rescale},
respectively.
\item \strong{startpos}: \code{numeric} gene start position in base
pairs. If the user is providing their own gene scores, then this
position represents the original position minus any buffer used to
assign summary statistics to genes. Required for gene score
deconfounding and gene set score decorrelation in and
\code{plR_permute} and \code{plR_rescale}, respectively.
\item \strong{endpos}: \code{numeric} gene end position in base pairs.
If the user is providing their own gene scores, then this position
represents the original position plus any buffer used to assign
summary statistics to genes. Required for gene score deconfounding
and gene set score decorrelation in and
\code{plR_permute} and \code{plR_rescale}, respectively.
\item \strong{startpos.base}: \code{numeric} gene start position in
base pairs. Default lower gene boundary (ignoring buffer used in gene
assignment). Created internally if absent or gene scores are
estimated (i.e. var.info provided). Required for gene score
deconfounding and set score decorrelation in \code{plR_permute}
and \code{plR_rescale}, respectively.
\item \strong{endpos.base}: \code{numeric} gene end position base
pairs. Default upper gene boundary (ignoring buffer used in gene
assignment). Created internally if absent or gene scores are
estimated (i.e. var.info provided). Required for gene score
deconfounding and set score decorrelation in \code{plR_permute}
and \code{plR_rescale}, respectively.
}
}
\item{\code{set.obj}}{
A \code{data.table} (and \code{data.frame}) mapping genes to gene sets.
\itemize{
\item \strong{setID}: \code{character} or \code{factor} vector of
unique gene set identifiers. Required.
\item \strong{objID}: \code{character} or \code{factor} vector of
unique gene identifiers. Required.
}
}
}
}

\section{Optional input file structure}{

These optional comma-separated (\code{.csv}) or tab-separated (\code{.tsv})
files provide the following \code{plR_read} functionality:

\describe{
\item{\code{var.info}}{
Contains information used to compute gene scores (\code{objStat}).
\itemize{
\item \strong{chr}: \code{character} or \code{numeric} chromosome /
contig labels. Required.
\item \strong{pos}: \code{numeric} position where statistic was
evaluated (e.g., SNP or central point in window). Required.
\item \strong{value}: \code{numeric} statistical value from SNP /
window score. Required.
}
}
\item{\code{rec.rate}}{
Used to transform genetic coordinates from physical to genetic
distances. Requires \code{startpos} and \code{endpos} in \code{obj.info}
to be base pair coordinates. Uses HapMap format (see labelled examples
at \url{https://zenodo.org/records/11437540}).
\itemize{
\item \strong{chr}: \code{character} or \code{numeric} chromosome /
contig labels. Required.
\item \strong{pos}: \code{numeric} base pair position of upstream
marker for the recombination interval. Required.
\item \strong{rate}: \code{numeric} recombination rate in cM per bp
in the downstream interval. Required.
\item \strong{map}: \code{numeric} genetic distance in centiMorgans
(cM). Optional; will be calculated from recombination rates if
absent.
}
}
}
}

\examples{
\dontrun{
# Example 1: Read all files from a single folder
my_plr <- read_polylinkr_data(input_path = "path/to/files")

# Example 2: Read all files with "POP1" in label from a single folder
my_plr <- read_polylinkr_data(
   input_path = "path/to/files",
   group_label = "POP1"
)

# Example 3: Relax gene set merging criteria
my_plr <- read_polylinkr_data(
   input_path = "path/to/files",
   merge_threshold = 0.50
)

# Example 4: Specify separate file paths and remove user-specified sets
my_plr <- read_polylinkr_data(
  gene_set_info_path = "path/to/set.info",
  gene_set_mapping_path = "path/to/set.obj",
  object_info_path = "path/to/obj.info",
  sets_to_exclude = c("set1", "set2")
)

# Example 5: Generate gene scores from variant info using regression
# and include a 50 kb buffer around genes
my_plr <- read_polylinkr_data(
  gene_set_info_path = "path/to/set.info",
  gene_set_mapping_path = "path/to/set.obj",
  object_info_path = "path/to/obj.info",
  variant_info_path = "path/to/var.info",
  object_statistic_function = "lm.logN",
  object_buffer = 5e4
)

# Example 6: Convert distances to cM using recombination rate file
my_plr <- read_polylinkr_data(
  gene_set_info_path = "path/to/set.info",
  gene_set_mapping_path = "path/to/set.obj",
  object_info_path = "path/to/obj.info",
  recombination_rate_path = "path/to/rec.rate"
)
}
}
