% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polylinkr.R
\name{ReadSetObjTables}
\alias{ReadSetObjTables}
\title{Read in polylink data.}
\usage{
ReadSetObjTables(
  in.path = "./",
  population = NA,
  minsetsize = 10,
  maxsetsize = 1000,
  merge.set.prop = 1,
  n.cores = NA
)
}
\arguments{
\item{in.path}{character: pathway to input files}

\item{population}{character: label used to identify input files. Three input files are required, specified by 'setInfo', 'setObj', and 'objInfo'.}

\item{minsetsize}{integer: minimum number of genes required in gene set}

\item{maxsetsize}{integer: maximum number of genes required in gene set}

\item{n.cores}{integer: number of cores to run in parallel}

\item{min.sim}{double: sets sharing at least this proportion of genes will be merged}
}
\value{
Returns a list with three elements.
}
\description{
Reads in data. Optionally, can merge gene sets that share more than a specified proportion of genes, or remove gene sets with less/more than a specified number of genes.
}
