% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale_polylinkr_data}
\alias{rescale_polylinkr_data}
\title{Gene set enrichment on scores rescaled for genetic autocorrelation}
\usage{
rescale_polylinkr_data(
  plr_input,
  rescale = TRUE,
  fast = TRUE,
  ac = NULL,
  cgm_bin = 30,
  cgm_range = "auto",
  cgm_weight_max = 0.05,
  empirical_bayes = "auto",
  min_rho = 1e-05,
  verbose = TRUE,
  n_cores = "auto",
  future_plan = "auto"
)
}
\arguments{
\item{plr_input}{\code{plR} class object; output from \code{polylinkR::permute_polylinkr_data}.
Required.}

\item{rescale}{\code{logical}; should gene set score (i.e.,\code{setScore.std}) rescaling
be performed? Defaults to \code{TRUE}. If \code{FALSE}, only inter-gene
autocorrelation is estimated without revising enrichment testing for
rescaled (i.e., decorrelated) gene set scores. This is useful for exploring
how parameter settings impact gene set autocorrelation. The rescaling step
can be performed later by passing the output to \code{rescale_polylinkr_data}, or by
providing the \code{ac} object to the \code{ac} argument.}

\item{fast}{\code{logical}; should the gene set score rescaling be performed using the
fast mode? Defaults to \code{TRUE}, whereby the analytical scaling factor
\eqn{1 / \sqrt{1 + (m - 1)\hat{\rho}}} is calculated from the covariance
matrix \code{ac} and used to adjust the empirical null and GPD (scale and
threshold) parameters. If \code{FALSE}, the full empirical rescaling is
performed instead, by recalculating the score for every permuted gene set.
This is much more computationally intensive and is intended for users
interested in comparing analytical and empirical results.}

\item{ac}{\code{data.frame} or \code{data.table}; user-provided object containing
inter-gene autocorrelations. Include all unique pairs of genes with
non-zero autocorrelation. Genes must be identified by \code{objID}s in
columns \code{objID.A} and \code{objID.B}. The corresponding
autocorrelation value must be in \code{gamma}. Defaults to \code{NULL}.}

\item{cgm_bin}{\code{numeric}; mimimum number of gene pairs required for bins of empirical
covariances. Default value = \code{30}. Must be in range \code{[10, 1e3]}.
Initially an exponential grid of bin sizes will be generated, favouring
smaller bins at short distances and larger bins for more distant gene pairs.
Bins with too few genes will be successively merged with the larger
bins until the minimum number of genes are reached. This condition is
evaluated across all chromosomes, ensuring no bin is smaller than the
minimum value (other than the final bins).}

\item{cgm_range}{\code{numeric}; maximum inter-gene lag used to evaluate autocovariance.
Defaults to \code{2e6} bp or \code{2} cM, depending on genetic distance
measure. Must be within interval \code{[1e5, 5e7]} bp or \code{[0.1, 50]}
cM.}

\item{cgm_weight_max}{\code{numeric}; maximum probability weight for a single lag window.
Defaults to \code{0.1}. Set to \code{1} if no upper bound is desired.
Note that the lower bound is reset to \code{2 / no. fitted lags} if the
chosen value falls below this limit.}

\item{empirical_bayes}{\code{numeric}; A character string indicating the empirical Bayes rescaling
framework employed for chromosome-level covariance beta coefficients. Users
may choose to fit either the \code{full} or \code{reduced} model, with the
the former model also allowing for the two coefficient to be correlated.
The default setting (\code{auto}) runs the \code{full} model if more than
\code{15} chromosomes are present, otherwise the \code{reduced} model is
fit. In cases when the \code{full} model is assessed, the  \code{reduced}
model is also fitted and a likelihood-based test is used to decide whether
a correlated random-effects structure is supported.}

\item{min_rho}{\code{numeric}; minimum estimated correlation between two gene sets;
values below this are set to \code{0}. Defaults to \code{1e-5}. Range
\code{(0, 0.01]}.}

\item{verbose}{\code{logical}; should progress reporting be enabled? Defaults to
\code{TRUE}.}

\item{n_cores}{\code{integer}; number of cores for parallel processing. Defaults to
\code{1} or \code{maximum cores - 1}. Must be in \code{[1, maximum cores]}.}

\item{future_plan}{\code{character}; parallel backend from the \code{future} package.
Defaults to user \code{n_cores} choice or checks cores, choosing
\code{"sequential"} on single-core and \code{"multisession"} on
multi-core systems. Options: \code{"multisession"},
\code{"multicore"}, \code{"cluster"}, \code{"sequential"}.}
}
\value{
A \code{plR} S3 object containing:
\describe{
\item{\code{obj.info}}{\code{data.table} for each gene (object).}
\item{\code{set.info}}{\code{data.table} for each gene set.}
\item{\code{set.obj}}{\code{data.table} mapping genes to gene sets.}
}

Rescaled set scores (i.e., corrected for correlations between genes within
sets) are recorded in \code{set.info} as \code{setScore.rs}. Results from
the revised enrichment tests are shown in \code{setScore.rs.p}.

All \code{plR} S3 objects include auxiliary information and summaries as
attributes:
\describe{
\item{\code{plr.data}}{
Reusable datasets and parameters, including GPD fitting results and
autocovariance estimation.
}
\item{\code{plr.args}}{
Argument settings used in the function.
}
\item{\code{plr.session}}{
R session information and function run time.
}
\item{\code{plr.track}}{
Internal tracking number indicating functional steps.
}
\item{\code{class}}{
S3 object class (i.e., \code{plr}).
}
}

Each attribute can be accessed using \code{attr()} or \code{attributes()}.
The first four attribute classes aggregate information over successive
functions. For example, to access the \code{plr.data} attribute for the
\code{plR} object output after running \code{plR_permute}, use
\code{attributes(X)$plR.data$permute.data}, where \code{X} is the object
name. Similarly, the arguments used in \code{plR_read} are in
\code{attributes(X)$plR.args$read.args}.

The primary data structure of the \code{plR} object can be accessed using
\code{print()} or by simply typing the object's name.
}
\description{
Rescales gene set scores to account for dependencies within gene sets based
on autocorrelated genes and re-evaluates gene set enrichment on the rescaled
scores. Note that the \code{obj.info} file must contain appropriately
labelled genomic coordinates for autocorrelation to be estimated and gene set
rescaling to be performed, otherwise the function will exit with an error
message.
}
\examples{
\dontrun{
# Assuming `my_plr` is the result of `polylinkR::permute_polylinkr_data`

# Example 1: Basic usage
new_plr <- rescale_polylinkr_data(plr_input = my_plr)

# Example 2: Estimate autocorrelation only (no rescaling)
new_plr <- rescale_polylinkr_data(
   plr_input = my_plr,
   rescale   = FALSE
)

# Example 3: Custom variogram model and cores
new_plr <- rescale_polylinkr_data(
  plr_input = my_plr,
  n_cores   = 4
)

# Example 4: Using a user-provided autocovariance object
# Assuming `my_ac` is a valid data.table or data.frame
new_plr <- rescale_polylinkr_data(
   plr_input = my_plr,
   ac        = my_ac
)

# Or using the `new_plr` object generated by Example 2
new_plr <- rescale_polylinkr_data(plr_input = my_rescaled_plr)
}
}
