% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anatolia_ef_clr}
\alias{anatolia_ef_clr}
\title{Anatolia gene expression example data}
\format{
A data frame with gene expression values.
}
\source{
Internal example dataset for testing.
}
\usage{
data(anatolia_ef_clr)
}
\description{
Example data: Gene expression factors from Anatolia population.
}
\keyword{datasets}
