% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{human_sexavg_RR_Bherer2017}
\alias{human_sexavg_RR_Bherer2017}
\title{human_sexavg_RR_Bherer2017 (deprecated)}
\format{
See \code{human_sexavg_rr_bherer2017}.
}
\description{
Deprecated alias for \code{human_sexavg_rr_bherer2017}.
}
\keyword{internal}
