% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{plR_prune}
\alias{plR_prune}
\title{Gene set enrichment adjusting for multiple testing and shared genes (deprecated)}
\usage{
plR_prune(
  plR.input,
  n.fdr = 300L,
  est.pi0 = TRUE,
  tolerance = 0.001,
  verbose = TRUE,
  n.cores = "auto",
  fut.plan = "auto"
)
}
\arguments{
\item{plR.input}{Deprecated. Use \code{plr_input}.}

\item{n.fdr}{Deprecated. Use \code{n_fdr}.}

\item{est.pi0}{Deprecated. Use \code{estimate_pi0}.}

\item{tolerance}{Deprecated. Use \code{tolerance}.}

\item{verbose}{Deprecated. Use \code{verbose}.}

\item{n.cores}{Deprecated. Use \code{n_cores}.}

\item{fut.plan}{Deprecated. Use \code{future_plan}.}
}
\description{
This function is deprecated. Please use \code{prune_polylinkr_data()} instead.
}
