% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plR-read.R
\name{plR_read}
\alias{plR_read}
\title{Read and validate required files for polylinkR (deprecated)}
\usage{
plR_read(...)
}
\arguments{
\item{...}{All arguments passed to \code{read_polylinkr_data()}.}
}
\description{
This function is deprecated. Please use \code{read_polylinkr_data()} instead.
}
\keyword{internal}
