% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polylinkR-package.R
\docType{package}
\name{polylinkR-package}
\alias{polylinkR}
\alias{polylinkR-package}
\title{polylinkR: Polygenic Pathway Enrichment with Linkage Structure}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Perform gene-based pathway enrichment analysis while accounting for linkage disequilibrium (LD) amongst loci. Uses a permutation approach that preserves linkage structure when generating null distributions of pathway scores. This reduces false positives compared to standard permutation methods. This version is under active development and must not be used as the sole basis for production or clinical decisions without independent validation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ACAD-UofA/PolyLinkR}
  \item Report bugs at \url{https://github.com/ACAD-UofA/PolyLinkR/issues}
}

}
\author{
\strong{Maintainer}: PolyLinkR authors \email{polylinkr@acad.edu.au}

Other contributors:
\itemize{
  \item ACAD University of Adelaide [copyright holder, funder]
}

}
\keyword{internal}
