% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{polylinkr_set_info}
\alias{polylinkr_set_info}
\title{polylinkr_set_info}
\format{
A data frame with setID, setName, setSource columns.
}
\source{
Internal example dataset for testing.
}
\description{
Example gene set information for testing.
}
