% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{polylinkr_set_obj}
\alias{polylinkr_set_obj}
\title{polylinkr_set_obj}
\format{
A data frame with setID and objID columns.
}
\source{
Internal example dataset for testing.
}
\description{
Example gene-to-set mapping for testing.
}
