% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{PolyLinkR_SetInfo}
\alias{PolyLinkR_SetInfo}
\title{PolyLinkR_SetInfo (deprecated)}
\format{
See \code{polylinkr_set_info}.
}
\description{
Deprecated alias for \code{polylinkr_set_info}.
}
\keyword{internal}
